Fix a read overflow at the end of y4m headers
For example, after the use of strtol, tokstart is set to one character
after the last of the number. The increment part of the for-loop increases
tokstart by one. If the parsed number was at the end of the header string,
tokstart is now one character after the '\0' at the end of the string. The
loop may then continue and fail because it tries to parse uninitialized
memory.

This commit transforms the for-loop into a while-loop without increment.
It also adds an assert to check that we do not try to parse further the
end of the header string.

This fixes random premature stops of the encoder with valid y4m inputs.