trivial: README.md: s/ran/run/
The past participle of the verb "to run" is "run", not "ran"; this
commit changes "ran" to "run".

Also, at the same time, this commit moves the preposition "on" from
the end of the sentence to a point earlier in the sentence.

The patch would be easiest to review with the following:

  git diff --word-diff