1pass BRC improvement
Adds support for 1-pass CBR real time mode and 1-pass vbr with the
ability to use LAD stats and calculate stat total in 1-pass VBR based on
look ahead. Also adds look ahead as an input for the encoder and unifies
the VBR algorithms between 2-pass and 1-pass. Creates ME functions for
IPPP processing of 2-pass and 1-pass VBR/CBR