fix early release of ppcs
issue:
"Error: PictureDecisionProcess: Picture Decision Reorder Queue overflow"
print and hang when enable overlay and lp != 1.
every input picture combines with an overlay ppcs when creating and
then releases if it is not alt-ref. An early release happens in PD and
recycles, then it is reused by other picture, error happens

fixing:
add live count +1 for PA and PD to avoid early release