fix issue 1919
root cause: size of input_frame16bit is changed to align
enhanced_downscaled_picture_ptr in super-res or resize enabled
frames, but it is not reset if frames in original size. a mismatch
happens between input_frame16bit and enhanced_picture_ptr in unscaled
frames.

fixing: add a reset of input_frame16bit to fix mismatch