Fix Neon compound convolution functions
Currently the Neon implementation of compound convolution functions uses
use_dist_wtd_comp_avg to determine the averaging mode. Change it to use
use_jnt_comp_avg instead to align with the unit tests and the other
implementations of the functions. Also add a test for the block size, as
the Neon implementations assume a minimal block size of 4x4.