Cleanup test names in CompoundUtil.cc
A lot of the naming in this file has redundant information between the
name of the test and the name of the instantiation. On the other hand,
it is not possible to filter tests for a single architecture variant.
Rename and split the tests so that one instantiation name corresponds to
one architecture extension. While we're here, rename all occurrences of
"Substract" to "Subtract".

Also take the opportunity to remove EbBlend_a64_mask_1d_test.cc and
EbBlend_a64_mask_test.cc as they are testing a subset of the same
functions tested in CompoundUtils.cc.