Port libaom implementation of highbd_convolve_2d_sr_neon
Replace the current Neon implementation of svt_av1_highbd_convolve_2d_sr
with the one present in libaom. This new version includes specialized
paths for each of the filter sizes. Also take the opportunity to move
the function to highbd_convolve_neon.cc with the rest of the convolution
functions.