Filter out unit tests for unsupported SIMD extensions
By default, the unit tests will try to run every variant of a given test
for the corresponding architecture (AArch64 or x86). However, this will
break if we try to run tests for architecture extensions unsupported by
the platform running the tests. To avoid this, create a main function
that will filter out tests at run time for SIMD extensions that are not
supported by the current execution platform. This will be useful when
Armv8.x and Armv9.x architecture feature paths are added.

All the logic is ported from the gtest main() function in libaom.