Resolve ABR first pass regression
This merge request resolves a regression with SVT-AV1 v2.2 where the speed of the first pass for ABR - two pass is significant reduced. This happened due to a reduction in the picture analysis processes.

To resolve this problem, the number of picture analysis processes is increased from 4 to 12 for the first pass of ABR - two pass only. The result is an uplift in performance of up to 80% for resolutions 1080p or lower.

Tested on i7-14700k.