Update libaom dependency
By default ExternalProject_Add() will re-download libaom every time
SVT-AV1 is built, which is unnecessary and makes the build far slower
than it should be. Add "UPDATE_DISCONNECTED True" to the call to
ExternalProject_Add() changes this behaviour so that libaom is only
downloaded once in the initial configure step. This means if libaom is
updated we will need to do a clean build of SVT-AV1. While we're here,
change the url to point to the actual libaom repo rather than a 5-year
old fork.