Enable LTO by default for Clang 12+
This merge request makes LTO the default build configuration for clang 12+. CMake implements Thin LTO for Clang. It is much less resource intensive at build time than GCC. It is also advantageous for performance at run time.

This Clang version was selected based on the SVT-AV1 project's ability to test and support the compiler versions. A total of 8 compiler versions are supported.