Fix unit tests for svt_av1_compute_cul_level
The tests input was such that the computed cul_level was always much
higher than the max return value (COEFF_CONTEXT_MASK = 63), therefore
the return value was always 63, regardless of what the function
actually does. As a result, both the AVX2 and NEON implementations were
computing the wrong value, but the unit tests were still passing. Change
the inputs to better reflect what happens during an actual encoding
(quant_coeff values are fairly low), so that the computed cul_level will
often be < 63.