feat: `--color-help` parameter for metadata
This MR enables a small quality of life CLI flag to print information in Appendix A.2 of the SVT-AV1 User Guide to help with setting color metadata options.
Note that the printed output uses ANSI escape codes, which are supported by most Linux & macOS systems in terminal emulators but may not be supported on Windows. To remedy this, we disable ANSI escape codes when SVT-AV1 is compiled to anything that looks like a Windows target.