Fix helper function of svt_av1_fwd_txfm2d_8x4_neon
The 8x4 transforms needs an extra multiply and rounding step due to the
width and height being different by a factor of 2, which is why we have
special TRANSFORM_ROW_RECT_* macros. However TRANSFORM_ROW_RECT_ONE was
missing this step which was "manually" added in the top-level function
instead. Move it back to inside the macro, and change the struct name to
match the selection.