Add more testcases for compute_cross_correlation unit tests
Test svt_av1_compute_cross_correlation with more values of match_sz. In
theory match_sz can be any odd number between 1 and 15, however if
match_sz = 1 then the variance is zero, so the function will return a
nan. Assuming this is not a case that happens in real life (the only
values set in enc_mode_config.c being 7 and 13), test every odd value
between 3 and 15.