Fix test coverage of CDEFSearchOneDualTest
Currently the unit test only runs the function with end_gi = 64,
while end_gi can take on the values {2, 3, 4, 6, 48, 64} - these
numbers were observed when running the end-to-end tests. The current
Neon implementation calls the scalar implementation if end_gi >= 48,
meaning the Neon implementation is not currently tested.

This change runs the unit test for more end_gi values, increasing
test coverage. It  also creates new test cases for even and odd values
of sb_count. (It was only tested for sb_count = 100 until now.)
This is required for a subsequent optimization patch that unrolls the loop.

Since the Neon version is now tested, call the Neon version of the
svt_memset function instead of the rtcd pointer, as these are not
automatically setup when running the unit tests, causing the tests to
segfault.