cmake: Support parallel shared and static installs
This allows the shared and static libraries to coexist if they are both
installed to the same prefix. In this case, the SVT-AV1::SVT-AV1 target
points to the shared library. Applications can choose the library they
want to link to by using the SVT-AV1::SVT-AV1-{shared,static} targets.
This will not work on Windows with MSVC though, due to the conflict
between import and static library names.